requirements.txt
flywheel-sdk>=15.10.0
pandas>=1.5.0
numpy>=1.22.0
SimpleITK>=2.2.0
nibabel>=5.0.0
matplotlib>=3.5.0
tqdm>=4.64.0
tabulate>=0.9.0
pathvalidate>=2.5.2
